refactor: migrate prioritypathogensview into sites/brc-analytics (#1554)
