+++ pages/data/priority-pathogens/index.tsx
@@ -5,7 +5,7 @@ import type {
 import { getEntityListMeta } from "@/common/meta/utils";
 import { config } from "@/config/config";
 import { seedDatabase } from "@/utils/seedDatabase";
-import { PriorityPathogensView } from "@/views/PriorityPathogensView/priorityPathogensView";
+import { PriorityPathogensView } from "@brc/views/PriorityPathogensView/priorityPathogensView";
 import { Main as DXMain } from "@databiosphere/findable-ui/lib/components/Layout/components/Main/main.styles";
 import { getEntityService } from "@databiosphere/findable-ui/lib/hooks/useEntityService";
 import { GetStaticProps } from "next";
