+++ tsconfig.json
@@ -21,6 +21,7 @@
       "@ga2/*": ["sites/ga2/*"],
       "@repo/shared": ["packages/shared/index.ts"],
       "@repo/shared/*": ["packages/shared/*"],
+      "@routes/*": ["routes/*"],
       "@site-config/*": ["site-config/*"],
       "images/*": ["images/*"]
     },
